+++ tests/Pipaslot.Mediator.Http.Tests/Serialization/ContractSerializerBaseTest.cs
