+++ tests/Pipaslot.Mediator.Http.Tests/Serialization/V3/JsonContractSerializer_CommonTests.cs
@@ -6,13 +6,12 @@ using Pipaslot.Mediator.Http.Serialization;
 using Pipaslot.Mediator.Http.Serialization.V3;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using Xunit;
 
 namespace Pipaslot.Mediator.Http.Tests.Serialization.V3
 {
     public class JsonContractSerializer_CommonTests : ContractSerializer_CommonTestBase
-    {      
+    {
 
         [Fact]
         public void DeserializeRequest_FromJsonWithShortFormat()
@@ -34,47 +33,7 @@ namespace Pipaslot.Mediator.Http.Tests.Serialization.V3
             var deserialized = sut.DeserializeResponse<PublicPropertyGetterAndInitSetterContract>(serialized);
 
             Assert.True(Match(deserialized.Result));
-        }
-
-        
-
-        [Fact]
-        public void Response_ResultTypeWithCollectionOfInterfaceInheritingCollection_WillKeepTheResultType()
-        {
-            var sut = CreateSerializer();
-            var result = new IsAuthorizedRequestResponse(new IRuleSet[] {
-                new RuleSet
-                {
-                    new Rule("name", "value", false)
-                }
-            });
-            var responseString = sut.SerializeResponse(new MediatorResponse(true, new object[] { result }));
-            var deserialized = sut.DeserializeResponse<IsAuthorizedRequestResponse>(responseString);
-            Assert.Equal(result.GetType(), deserialized.Result.GetType());
-        }
-
-        [Fact]
-        public void Response_ResultTypeWithCollectionOfObjectInheritingCollection_WillKeepTheResultType()
-        {
-            var sut = CreateSerializer();
-            var result = new FakeRuleSetWithSpecificArrayResponse
-            {
-                NotMetRule = new RuleSet[] {
-                    new RuleSet
-                    {
-                        new Rule("name", "value", false)
-                    }
-                }
-            };
-            var responseString = sut.SerializeResponse(new MediatorResponse(true, new object[] { result }));
-            var deserialized = sut.DeserializeResponse<FakeRuleSetWithSpecificArrayResponse>(responseString);
-            Assert.Equal(result.GetType(), deserialized.Result.GetType());
-        }
-        public class FakeRuleSetWithSpecificArrayResponse
-        {
-            public bool IsAuthorized => NotMetRule.Count() == 0;
-            public RuleSet[] NotMetRule { get; set; } = Array.Empty<RuleSet>();
-        }
+        }       
 
         public class MessageWithInterfaceProperty : IMessage
         {
@@ -93,10 +52,6 @@ namespace Pipaslot.Mediator.Http.Tests.Serialization.V3
         public class Contract : IContract
         {
             public string Name { get; set; } = string.Empty;
-        }
-        public class MessageWithIMediatorActionProperty : IMessage
-        {
-            public IMediatorAction SubAction { get; set; } = null!;
         }
         public class ChildMediatorAction : IMessage
         {
